change project arhitecture
